custom list delete insert part done

+++ app.js
@@ -35,7 +35,7 @@ app.use(express.static("public"));
 app.get("/", (req, res) => {
     Item.find({}, async (err, result) => {
         if (!err) {
-            res.render('index', { Data: result, Title: "MY TO Do List" })
+            res.render('index', { Data: result, Title: "MY To Do List" })
         }
     })
 
@@ -44,32 +44,55 @@ app.post("/", (req, res) => {
     const item = req.body.item;
     const Title = req.body.Title;
     const data = new Item({ name: item })
+    if(Title=="MY To Do List"){
         data.save();
-    res.redirect('/')
+        res.redirect('/')
+    }else{
+        Custom.findOne({name:Title},(err,result)=>{
+         result.item.push(data);
+         result.save();
+         res.redirect("/"+Title);
+        })
+    }
+       
 })
 
 app.post("/delete", (req, res) => {
-    const ItemId = req.body.ListId
-    Item.findByIdAndRemove(ItemId, (err) => {
-        if (!err) {
-            console.log("successfully deleted");
-        }
-    })
-    res.redirect('/')
+    const ItemId = req.body.ListId;
+     const Title = req.body.Title;
+    
+     if(Title=="My To Do List"){
+        Item.findByIdAndRemove(ItemId, (err) => {
+            if (!err) {
+                console.log("successfully deleted");
+            }
+        })
+        res.redirect('/');
+     }else{
+        Custom.findOne({name:Title},(err,result)=>{
+            if(!err){
+                result.item.pull({_id:ItemId});
+                result.save();
+                res.redirect("/"+Title);
+            }
+        })
+     }
+   
 })
 
 app.get("/:customTitle", (req, res) => {
     const customTitle = req.params.customTitle;
 
-    Custom.findOne({ name: customTitle }, (err, result) => {
+    Custom.findOne({name:customTitle}, (err, result) => {
         if (!err) {
-          if(!result){
-            const newItem = new Custom({name:Custom});
+         if(!result){
+            const newItem =new Custom({name:customTitle})
             newItem.save();
-            res.redirect("/"+customTitle);
-          }else{
-            res.render('index', { Title: customTitle, Data: result.item })
-          }
+            res.redirect("/"+customTitle)
+         }else{
+            res.render('index', { Title: customTitle, Data: result.item })           
+         }
+           
         }
        
     })
